==> growth_curve/03-Missings.Rmd <==
---
title: '3 Visualization of Missing Data'
author: 
  - name: "Júlia Cano Raluy"
    affiliation: "Mathematics student at Autonomous University of Barcelona (UAB)"
  - name: "Juan R González"
    affiliation: 
      - "Bioinformatics Research Group in Epidemiology (BRGE), Barcelona Institute for Global Health (ISGlobal)"
      - "Department of Mathematics, Autonomous University of Barcelona (UAB)"
    email: juanr.gonzalez@isglobal.org
date: "`r Sys.Date()`"
output: html_document
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE)

packages <- c("tidyverse","summarytools","caret","reshape2","ggplot2","recipes","factoextra","magick","lubridate","cowplot","zoo","mice",
              "VIM","lattice","MVN","mvoutlier","gridExtra", "lattice")

install_and_load <- function(package) {
  if (!require(package, character.only = TRUE)) {
    install.packages(package, dependencies = TRUE)
    library(package, character.only = TRUE)
  }
}

invisible(lapply(packages, install_and_load))

load("densitometriapre2.Rdata")
load("ecocardiogramapre2.Rdata")
```


##  Contents

1. [Visualization of Missing Data](#visualization-of-missing-data)
2. [Multiple Imputation of Missing Continuous Variables](#multiple-imputation-of-missing-continuous-variables)
  1. [Validation of Imputation](#validation-of-imputation)
3. [Imputation of Categorical Variables](#imputation-of-categorical-variables)
  1.[Categorization of Age](#categorization-of-age)


# Visualization of Missing Data 

We will start by visualizing the amount of missing data in each dataframe 

```{r, warning=FALSE} 
missings_eco <- dfpreeco[, colSums(is.na(dfpreeco)) != 0 ] %>% gather(key = "variable", value = "valor")
missings_desi <- dfpredensi[, colSums(is.na(dfpredensi)) != 0 ] %>%  gather(key = "variable", value = "valor")
```

```{r,  fig.cap=c("Valores perduts ecocardiogrames", "Valors perduts desitometries"), fig.width=7, fig.height=8}
# Echocardiogram
missings_summary <- missings_eco %>%
  group_by(variable) %>%
  dplyr::summarize(percentatges_NA = 100 * sum(is.na(valor)) / length(valor)) %>%
  arrange(desc(percentatges_NA)) 

ggplot(data = missings_summary, aes(x = percentatges_NA, y = reorder(variable, -percentatges_NA))) +
  geom_col() +
  labs(title = "Percentatge valors absents per variable",
       x = "Percentatge NAs", y = "Variable") +
  theme_bw() +
  theme(axis.text.y = element_text(angle = 0, hjust = 1))



# Densitometry
missings_summary <- missings_desi %>%
  group_by(variable) %>%
  dplyr::summarize(percentatges_NA = 100 * sum(is.na(valor)) / length(valor)) %>%
  arrange(desc(percentatges_NA))  

ggplot(data = missings_summary, aes(x = reorder(variable, -percentatges_NA), y = percentatges_NA)) +
  geom_col() +
  labs(title = "Percentatge valors absents per variable",
       x = "Variable", y = "Percentatge NAs") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1))+
  coord_flip()

# Columns without NAs
names(dfpredensi[, colSums(is.na(dfpredensi)) == 0 ])
```

# Multiple Imputation of Missing Continuous Variables

We will utilize the MICE library to perform multiple imputation for continuous variables, generating multiple datasets with various imputations. The quantitative variables will be imputed using the Predictive Mean Matching (PMM) method.

```{r} 
set.seed(123)
dfpreeco1 <- dfpreeco %>%
  dplyr::select(where(is.numeric))
dfpredensi1 <- dfpredensi %>%
  dplyr::select(where(is.numeric))

initeco = mice::mice(dfpreeco1, maxit = 0)
initdensi = mice::mice(dfpredensi1, maxit = 0)
```

```{r, results='hide'}
set.seed(123)
imputedeco = mice(dfpreeco1, method=initeco$method, predictorMatrix= initeco$predictorMatrix, m=30) 
set.seed(123)
imputeddensi = mice(dfpredensi1, method=initdensi$method, predictorMatrix= initdensi$predictorMatrix, m=30)
save(imputedeco, file = "imputedeco.Rdata")
save(imputeddensi, file = "imputedensi.Rdata")
```

## Validation of Imputation
Before replacing the imputed values in our datasets, let's check the imputations.
We will use the density distributions of the initial values (blue) and the different imputations (red). We observe that they do not change much between them and are similar to the original values.

```{r, warning=FALSE}
set.seed(123)
imputedeco10 = mice(dfpreeco1, method=initeco$method, predictorMatrix= initeco$predictorMatrix, m=10) 
densityplot(imputedeco10)
```

For each variable, we will obtain a dataframe with 30 columns of imputations and as many rows as NAs need to be imputed in that variable. To merge them into a single column, we will take the mean and replace it in the initial data.

```{r}
imputedeco$imp <- map(map(imputedeco$imp, rowMeans), ~ data.frame(.x))
imputeddesi$imp <- map(map(imputeddesi$imp, rowMeans), ~ data.frame(.x))

colseco <-setdiff(names(dfpreeco), names(dfpreeco1))
colsdensi <- setdiff(names(dfpredensi), names(dfpredensi1))

dfpreeco <- mice::complete(imputedeco, 1) %>% bind_cols(dfpreeco[colseco]) 
dfpredensi<- mice::complete(imputeddesi, 1) %>% bind_cols(dfpredensi[colsdensi])
```

# Imputation of Categorical Variables

Finally, we need to impute the categorical variables. We are missing the columns Section, Sex, Test, Date, and Year. For the variables Sex, Test, Date, and Year, we do not have missing values.
Therefore, only Section is missing. For now, we will delete the rows with missing values and save the IDs without Section.
```{r}
naseccioeco <- dfpreeco  %>%
    filter(is.na(Seccio)) %>% 
    distinct(Id)

nasecciodensi <-dfpredensi %>% 
  filter(is.na(Seccio)) %>% 
  distinct(Id)

NAseccio <- rbind(nasecciodensi, naseccioeco) %>%  distinct(Id)
save(NAseccio, file = "NAseccio.Rdata") #Id sense Secció

dfpreeco <- na.omit(dfpreeco, subset = Seccio)
dfpredensi <- na.omit(dfpredensi, subset = Seccio)
```

Now that we don't have missing values in the age variables, we can categorize them by terciles.

```{r}
set.seed(123)
quantile(c(dfpreeco$Edat, dfpredensi$Edat), probs = c(1/3, 2/3), na.rm = TRUE)
min(c(dfpreeco$Edat, dfpredensi$Edat), na.rm = TRUE)
max(c(dfpreeco$Edat, dfpredensi$Edat), na.rm = TRUE)


dfpreeco$Tercil_Edat <- cut(dfpreeco$Edat, 3, labels = c("Joves", "Aduls_Joves", "Adults"))
dfpredensi$Tercil_Edat <- cut(dfpredensi$Edat, 3, labels = c("Joves", "Aduls_Joves", "Adults"))

```
Thus, the ages will be classified into three groups: Young, Young Adults, and Adults.

```{r, include= FALSE}
save(dfpreeco, file = "densitometriapre3.Rdata")
save(dfpredensi, file = "ecocardiogramapre3.Rdata")
objetos <- Filter(function(x) x != "dfpreeco" & x != "dfpredensi", ls())
rm(list = objetos)
```
